Fix registrations cannot be saved via the admin
The event question multiple choice field (Anwer.option) is optional in a registration form.
Whether it is a required field is defined in the related question.
However, it is always visible in the admin but did not have blank=True defined. This prevents saving the a registration in the admin if there is no value selected.

Fixed by adding blank=True to the Anwer.option field.